+++ Axi4/TestCases/TbAxi4_AxiMemoryRandomTiming1.vhd
@@ -1,5 +1,5 @@
 --
---  File Name:         TbAxi4_MemoryRandomTiming1.vhd
+--  File Name:         TbAxi4_AxiMemoryRandomTiming1.vhd
 --  Design Unit Name:  Architecture of TestCtrl
 --  Revision:          OSVVM MODELS STANDARD VERSION
 --
@@ -9,7 +9,7 @@
 --
 --
 --  Description:
---      Testing of Burst Features in AXI Model
+--      Testing of Burst Features in AxiMemory Model
 --
 --
 --  Developed by:
@@ -19,12 +19,12 @@
 --
 --  Revision History:
 --    Date      Version    Description
---    05/2023   2023.05    Initial revision
+--    04/2025   2025.04    Initial revision
 --
 --
 --  This file is part of OSVVM.
 --
---  Copyright (c) 2023 by SynthWorks Design Inc.
+--  Copyright (c) 2025 by SynthWorks Design Inc.
 --
 --  Licensed under the Apache License, Version 2.0 (the "License");
 --  you may not use this file except in compliance with the License.
@@ -39,7 +39,7 @@
 --  limitations under the License.
 --
 
-architecture MemoryRandomTiming1 of TestCtrl is
+architecture AxiMemoryRandomTiming1 of TestCtrl is
 
   signal SyncPoint, TestDone, WriteDone : integer_barrier := 1 ;
   constant BURST_MODE : AddressBusFifoBurstModeType := ADDRESS_BUS_BURST_WORD_MODE ;
@@ -55,7 +55,7 @@ begin
   ControlProc : process
   begin
     -- Initialization of test
-    SetTestName("TbAxi4_MemoryRandomTiming1") ;
+    SetTestName("TbAxi4_AxiMemoryRandomTiming1") ;
     SetLogEnable(PASSED, TRUE) ;   -- Enable PASSED logs
     SetLogEnable(INFO, TRUE) ;     -- Enable INFO logs
     -- SetLogEnable(DEBUG, TRUE) ;    -- Enable DEBUG logs
@@ -74,7 +74,7 @@ begin
 
     TranscriptClose ;
     -- Printing differs in different simulators due to differences in process order execution
-    -- AlertIfDiff("./results/TbAxi4_MemoryRandomTiming1.txt", "../AXI4/Axi4/testbench/validated_results/TbAxi4_MemoryRandomTiming1.txt", "") ;
+    -- AffirmIfTranscriptsMatch("../AXI4/Axi4/testbench/validated_results") ;
 
     EndOfTestReports(TimeOut => (now >= 10 ms)) ;
     std.env.stop ;
@@ -86,13 +86,6 @@ begin
   --   Generate transactions for AxiManager
   ------------------------------------------------------------
   ManagerProc : process
-    variable BurstVal  : AddressBusFifoBurstModeType ;
-    variable RxData    : std_logic_vector(31 downto 0) ;
-    constant DATA_ZERO : std_logic_vector := (DATA_WIDTH - 1 downto 0 => '0') ;
-    variable CoverID1, CoverID2 : CoverageIdType ;
-    variable slvBurstVector : slv_vector(1 to 5)(31 downto 0) ;
-    variable intBurstVector : integer_vector(1 to 5) ;
-    variable DelayCov       : AxiDelayCoverageIdArrayType ; 
   begin
     wait until nReset = '1' ;
     WaitForClock(ManagerRec, 2) ;
@@ -144,12 +137,44 @@ begin
   --   Generate transactions for AxiSubordinate
   ------------------------------------------------------------
   MemoryProc : process
-    variable Addr : std_logic_vector(AXI_ADDR_WIDTH-1 downto 0) ;
-    variable Data : std_logic_vector(AXI_DATA_WIDTH-1 downto 0) ;
-    variable DelayCov       : AxiDelayCoverageIdArrayType ; 
+    variable DelayCovID       : AxiDelayCoverageIdArrayType ; 
   begin
+    -- Use Coverage based delays
     SetUseRandomDelays(SubordinateRec) ; 
-    WaitForClock(SubordinateRec, 2) ;
+
+    -- Testing AXI Specific Random Delay Capability
+    GetDelayCoverageID(SubordinateRec, DelayCovID) ; 
+
+    -- Change all of the Coverage Bins
+    DeallocateBins(DelayCovID(WRITE_ADDRESS_ID)) ;  -- Remove old coverage model
+    AddBins (DelayCovID(WRITE_ADDRESS_ID).BurstLengthCov,    GenBin(7)) ; 
+    AddCross(DelayCovID(WRITE_ADDRESS_ID).BurstDelayCov,     GenBin(1), GenBin(2)) ;
+    AddCross(DelayCovID(WRITE_ADDRESS_ID).BeatDelayCov,      GenBin(0), GenBin(0)) ;
+
+    DeallocateBins(DelayCovID(WRITE_DATA_ID)) ;  -- Remove old coverage model
+    AddBins (DelayCovID(WRITE_DATA_ID).BurstLengthCov,       GenBin(6)) ; 
+    AddCross(DelayCovID(WRITE_DATA_ID).BurstDelayCov,        GenBin(1), GenBin(4)) ;
+    AddCross(DelayCovID(WRITE_DATA_ID).BeatDelayCov,         GenBin(0), GenBin(0)) ;
+
+    DeallocateBins(DelayCovID(WRITE_RESPONSE_ID)) ;  -- Remove old coverage model
+    AddBins (DelayCovID(WRITE_RESPONSE_ID).BurstLengthCov,  GenBin(4)) ;
+    AddBins (DelayCovID(WRITE_RESPONSE_ID).BurstDelayCov,   GenBin(3)) ;
+    AddBins (DelayCovID(WRITE_RESPONSE_ID).BeatDelayCov,    GenBin(0)) ;
+
+    DeallocateBins(DelayCovID(READ_ADDRESS_ID)) ;  -- Remove old coverage model
+    AddBins (DelayCovID(READ_ADDRESS_ID).BurstLengthCov,     GenBin(8)) ; 
+    AddCross(DelayCovID(READ_ADDRESS_ID).BurstDelayCov,      GenBin(1), GenBin(3)) ;
+    AddCross(DelayCovID(READ_ADDRESS_ID).BeatDelayCov,       GenBin(0), GenBin(0)) ;
+
+    DeallocateBins(DelayCovID(READ_DATA_ID)) ;  -- Remove old coverage model
+    AddBins (DelayCovID(READ_DATA_ID).BurstLengthCov,       GenBin(6)) ;
+    AddBins (DelayCovID(READ_DATA_ID).BurstDelayCov,        GenBin(5)) ;
+    AddBins (DelayCovID(READ_DATA_ID).BeatDelayCov,         GenBin(0)) ;
+
+    --
+    -- At this point, the Axi4Memory responds to the transactions from the 
+    -- Axi4Manager using the updated DelayCoverage models above
+    --
 
     -- Wait for outputs to propagate and signal TestDone
     WaitForClock(SubordinateRec, 2) ;
@@ -158,12 +183,12 @@ begin
   end process MemoryProc ;
 
 
-end MemoryRandomTiming1 ;
+end AxiMemoryRandomTiming1 ;
 
-Configuration TbAxi4_MemoryRandomTiming1 of TbAxi4Memory is
+Configuration TbAxi4_AxiMemoryRandomTiming1 of TbAxi4Memory is
   for TestHarness
     for TestCtrl_1 : TestCtrl
-      use entity work.TestCtrl(MemoryRandomTiming1) ;
+      use entity work.TestCtrl(AxiMemoryRandomTiming1) ;
     end for ;
   end for ;
-end TbAxi4_MemoryRandomTiming1 ;
+end TbAxi4_AxiMemoryRandomTiming1 ;